Refactor version parsing for testability
