+++ src/input.ts
@@ -1,3 +1,4 @@
+import {CalendarVersion, Version, parseVersion} from './version'
 import moment from 'moment'
 
 const commas = /\s*,\s*/
@@ -11,4 +12,20 @@ export function parseMonths(lines: string[]): number[] {
     .filter(m => !isNaN(m))
 
   return [...new Set(months)].sort((a, b) => a - b)
+}
+
+function notNull<T>(v: T | null): v is T {
+  return v !== null
+}
+
+export function parseLatestVersion(
+  prefix: string,
+  tags: string[]
+): CalendarVersion {
+  return tags
+    .filter(t => t.startsWith(prefix))
+    .map(t => t.replace(prefix, ''))
+    .map(parseVersion)
+    .filter(notNull)
+    .reduce((a, b) => (a.compare(b) > 0 ? a : b), new Version(0, 0))
 }
